Pure refactoring to merge domain_worker and domain_worker_starter into one
